+++ pybind/pymodel.cpp
@@ -305,12 +305,10 @@ void PyModel::single_evo_details(Shock const& shock, Coord const& coord, Observe
 
     if (rad.ssc) {
         if (rad.kn) {
-            KN_cooling(syn_e, syn_ph, shock, coord, obs_setup.z);
+            KN_cooling(syn_e, syn_ph, shock, coord);
         } else {
-            Thomson_cooling(syn_e, syn_ph, shock, coord, obs_setup.z);
+            Thomson_cooling(syn_e, syn_ph, shock, coord);
         }
-    } else if (rad.rad.cmb_cooling) {
-        CMB_cooling(syn_e, syn_ph, shock, coord, obs_setup.z);
     }
     save_electron_details(syn_e, details);
     save_photon_details(syn_ph, details, shock);
@@ -544,12 +542,10 @@ auto PyModel::sky_image(PyArray const& t_obs, double nu_obs, double fov, size_t 
 
         if (rad.ssc) {
             if (rad.kn) {
-                KN_cooling(syn_e, syn_ph, shock, coord, obs_setup.z);
+                KN_cooling(syn_e, syn_ph, shock, coord);
             } else {
-                Thomson_cooling(syn_e, syn_ph, shock, coord, obs_setup.z);
+                Thomson_cooling(syn_e, syn_ph, shock, coord);
             }
-        } else if (rad.rad.cmb_cooling) {
-            CMB_cooling(syn_e, syn_ph, shock, coord, obs_setup.z);
         }
 
         auto img = observer.sky_image(coord, shock, t_arr, nu_cgs, syn_ph, npixel, pix_size);